+++ tests/CommBiz.Api.Tests/Host/HealthEndpointTests.cs
@@ -1,7 +1,7 @@
 using System.Net;
 using Microsoft.AspNetCore.Mvc.Testing;
 
-namespace CommBiz.Api.Tests;
+namespace CommBiz.Api.Tests.Host;
 
 public class HealthEndpointTests(WebApplicationFactory<Program> factory) : IClassFixture<WebApplicationFactory<Program>>
 {